reworked ship movement + minor fixes
simplified original ship movement algorithm
holding shift locks ship position
holding control disables attack move

+++ StarshipArena/src/MachineGunPod.java
@@ -20,11 +20,11 @@ public class MachineGunPod extends BasicPod{
 		super(mygame, spawnx, spawny);
 	}
 	
-	public MachineGunPod(StarshipArena mygame, int spawnx, int spawny, int spawnangle){
+	public MachineGunPod(StarshipArena mygame, int spawnx, int spawny, double spawnangle){
 		super(mygame, "none", spawnx, spawny, spawnangle);
 	}
 
-	public MachineGunPod(StarshipArena mygame, String newteam, int spawnx, int spawny, int spawnangle){
+	public MachineGunPod(StarshipArena mygame, String newteam, int spawnx, int spawny, double spawnangle){
 		super(mygame, newteam, spawnx, spawny, spawnangle);
 	}
